requirements.txt
fastapi
uvicorn
python-multipart
requests
numpy
onnxruntime
Pillow
opencv-python-headless
jinja2
aiofiles